Moved useMediaQuery outside of component
